Work around the communication <server-car>, and other stuff (1/3)

+++ ManufacturerWebServer/src/main/java/sirs/motorist/prototype/controller/UserConfigController.java
@@ -5,6 +5,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import pt.tecnico.sirs.secdoc.Check;
 import sirs.motorist.prototype.model.dto.ConfigurationIdRequestDto;
 import sirs.motorist.prototype.model.dto.UserPairRequestDto;
 import sirs.motorist.prototype.model.entity.Configuration;
@@ -17,20 +18,30 @@ public class UserConfigController {
     private static final Logger logger = LoggerFactory.getLogger(UserConfigController.class);
 
     UserConfigService userConfigService;
+    Check check;
 
     @Autowired
     public UserConfigController(UserConfigService userConfigService) {
         this.userConfigService = userConfigService;
+        this.check = nonceChecker;
     }
 
     @PostMapping("/pair")
     public ResponseEntity<?> pairNewUser(@RequestBody UserPairRequestDto request) {
-        //TODO: to implement
-        return ResponseEntity.ok(":)");
+        if(check.verifyNonce(request.getNonce())) {
+            return ResponseEntity.badRequest().body("Nonce verification failed");
+        }
+        if(!userConfigService.pairNewUser(request)) {
+            return ResponseEntity.badRequest().body("Error pairing new user");
+        }
+        return ResponseEntity.ok("User paired successfully");
     }
 
     @PostMapping("/readConfig")
-    public ResponseEntity<?> readCurrentConfig(@RequestBody ConfigurationIdRequestDto request) { // TODO: BIG MAYBE to use cookies in the future
+    public ResponseEntity<?> readCurrentConfig(@RequestBody ConfigurationIdRequestDto request) {
+        if(check.verifyNonce(request.getNonce())) {
+            return ResponseEntity.badRequest().body("Nonce verification failed");
+        }
         Configuration response = userConfigService.getConfiguration(request.getUserId(), request.getCarId());
         if (response == null) {
             logger.error("Configuration for that user and car was not found...");
@@ -41,6 +52,9 @@ public class UserConfigController {
 
     @PutMapping("/updateConfig")
     public ResponseEntity<?> updateConfig(@RequestBody Configuration request) {
+        if(check.verifyNonce(request.getNonce())) {
+            return ResponseEntity.badRequest().body("Nonce verification failed");
+        }
         if(!userConfigService.updateConfiguration(request)) {
             return ResponseEntity.badRequest().body("Error updating configuration");
         }
@@ -49,6 +63,9 @@ public class UserConfigController {
 
     @PutMapping("/deleteConfig")
     public ResponseEntity<?> deleteConfig(@RequestBody ConfigurationIdRequestDto request) {
+        if(check.verifyNonce(request.getNonce())) {
+            return ResponseEntity.badRequest().body("Nonce verification failed");
+        }
         if(!userConfigService.deleteConfiguration(request)) {
             return ResponseEntity.badRequest().body("Error resetting the configuration");
         }
